refactor(desktop): split renderer app shell entry
